Move shared logic to base call handler
